+++ mediation/IronSource/source/plugin/Assets/GoogleMobileAds/Common/Mediation/IronSource/IIronSourceClient.cs
@@ -19,5 +19,6 @@ namespace GoogleMobileAds.Common.Mediation.IronSource
     public interface IIronSourceClient
     {
         void SetConsent(bool consent);
+        void SetMetaData(string key, string metaDataValue);
     }
 }